Move androidMain package to commonMain
