--- sample/src/androidMain/kotlin/com/davidmigloz/numberkeyboard/sample/Platform.android.kt
@@ -1,9 +1,0 @@
-package com.davidmigloz.numberkeyboard.sample
-
-import android.os.Build
-
-class AndroidPlatform : Platform {
-    override val name: String = "Android ${Build.VERSION.SDK_INT}"
-}
-
-actual fun getPlatform(): Platform = AndroidPlatform()
