+++ sample/src/commonMain/kotlin/com/davidmigloz/numberkeyboard/sample/App.kt
@@ -1,43 +1,90 @@
 package com.davidmigloz.numberkeyboard.sample
 
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.safeContentPadding
-import androidx.compose.material3.Button
-import androidx.compose.material3.MaterialTheme
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.rounded.AttachMoney
+import androidx.compose.material.icons.rounded.DashboardCustomize
+import androidx.compose.material.icons.rounded.Fingerprint
+import androidx.compose.material.icons.rounded.Numbers
+import androidx.compose.material3.Icon
+import androidx.compose.material3.NavigationBar
+import androidx.compose.material3.NavigationBarItem
+import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.navigation.NavController
+import androidx.navigation.NavDestination.Companion.hierarchy
+import androidx.navigation.NavGraph.Companion.findStartDestination
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.currentBackStackEntryAsState
+import androidx.navigation.compose.rememberNavController
+import com.davidmigloz.numberkeyboard.sample.theme.AppTheme
 import org.jetbrains.compose.ui.tooling.preview.Preview
 
 @Composable
 @Preview
 fun App() {
-    MaterialTheme {
-        var showContent by remember { mutableStateOf(false) }
-        Column(
-            modifier = Modifier
-                .safeContentPadding()
-                .fillMaxSize(),
-            horizontalAlignment = Alignment.CenterHorizontally,
-        ) {
-            Button(onClick = { showContent = !showContent }) {
-                Text("Click me!")
-            }
-            AnimatedVisibility(showContent) {
-                val greeting = remember { Greeting().greet() }
-                Column(Modifier.fillMaxWidth(), horizontalAlignment = Alignment.CenterHorizontally) {
-//                    Image(painterResource(Res.drawable.compose_multiplatform), null)
-                    Text("Compose: $greeting")
-                }
+    AppTheme {
+        val navController = rememberNavController()
+        Scaffold(bottomBar = { NavigationBar(navController) }) { innerPadding ->
+            NavHost(navController = navController, startDestination = BottomNavItem.Integer.route) {
+                composable(BottomNavItem.Integer.route) { IntegerScreen(innerPadding) }
+                composable(BottomNavItem.Decimal.route) { DecimalScreen(innerPadding) }
+                composable(BottomNavItem.Biometric.route) { BiometricScreen(innerPadding) }
+                composable(BottomNavItem.Custom.route) { CustomScreen(innerPadding) }
             }
         }
     }
+}
+
+@Composable
+private fun NavigationBar(navController: NavController) {
+    val navBackStackEntry by navController.currentBackStackEntryAsState()
+    val currentDestination = navBackStackEntry?.destination
+
+    NavigationBar {
+        BottomNavItem.entries.forEach { screen ->
+            NavigationBarItem(
+                icon = { Icon(screen.icon, contentDescription = screen.title) },
+                label = { Text(screen.title) },
+                selected = currentDestination?.hierarchy?.any { it.route == screen.route } == true,
+                onClick = {
+                    navController.navigate(screen.route) {
+                        popUpTo(navController.graph.findStartDestination().id) { saveState = true }
+                        launchSingleTop = true
+                        restoreState = true
+                    }
+                }
+            )
+        }
+    }
+}
+
+enum class BottomNavItem(
+    val title: String,
+    val route: String,
+    val icon: ImageVector,
+) {
+    Integer(
+        "Integer",
+        "integer",
+        Icons.Rounded.Numbers
+    ),
+    Decimal(
+        "Decimal",
+        "decimal",
+        Icons.Rounded.AttachMoney
+    ),
+    Biometric(
+        "Biometric",
+        "biometric",
+        Icons.Rounded.Fingerprint
+    ),
+    Custom(
+        "Custom",
+        "custom",
+        Icons.Rounded.DashboardCustomize
+    )
 }
