+++ sample/src/commonMain/kotlin/com/davidmigloz/numberkeyboard/sample/DecimalScreen.kt
