--- sample/src/commonMain/kotlin/com/davidmigloz/numberkeyboard/sample/Greeting.kt
@@ -1,9 +1,0 @@
-package com.davidmigloz.numberkeyboard.sample
-
-class Greeting {
-    private val platform = getPlatform()
-
-    fun greet(): String {
-        return "Hello, ${platform.name}!"
-    }
-}
