+++ sample/src/commonMain/kotlin/com/davidmigloz/numberkeyboard/sample/IntegerScreen.kt
