--- sample/src/commonMain/kotlin/com/davidmigloz/numberkeyboard/sample/Platform.kt
@@ -1,7 +1,0 @@
-package com.davidmigloz.numberkeyboard.sample
-
-interface Platform {
-    val name: String
-}
-
-expect fun getPlatform(): Platform
