--- sample/src/desktopMain/kotlin/com/davidmigloz/numberkeyboard/sample/Platform.jvm.kt
@@ -1,7 +1,0 @@
-package com.davidmigloz.numberkeyboard.sample
-
-class JVMPlatform: Platform {
-    override val name: String = "Java ${System.getProperty("java.version")}"
-}
-
-actual fun getPlatform(): Platform = JVMPlatform()
