--- sample/src/iosMain/kotlin/com/davidmigloz/numberkeyboard/sample/Platform.ios.kt
@@ -1,9 +1,0 @@
-package com.davidmigloz.numberkeyboard.sample
-
-import platform.UIKit.UIDevice
-
-class IOSPlatform: Platform {
-    override val name: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
-}
-
-actual fun getPlatform(): Platform = IOSPlatform()
